# [One .NET] simplify native library inclusion