… src/Xamarin.Android.Build.Tasks/Properties/xlf/Resources.cs.xlf
… src/Xamarin.Android.Build.Tasks/Properties/xlf/Resources.de.xlf
… src/Xamarin.Android.Build.Tasks/Properties/xlf/Resources.es.xlf
… src/Xamarin.Android.Build.Tasks/Properties/xlf/Resources.fr.xlf
… src/Xamarin.Android.Build.Tasks/Properties/xlf/Resources.it.xlf
… src/Xamarin.Android.Build.Tasks/Properties/xlf/Resources.ja.xlf
… src/Xamarin.Android.Build.Tasks/Properties/xlf/Resources.ko.xlf
… src/Xamarin.Android.Build.Tasks/Properties/xlf/Resources.pl.xlf
… src/Xamarin.Android.Build.Tasks/Properties/xlf/Resources.pt-BR.xlf
… src/Xamarin.Android.Build.Tasks/Properties/xlf/Resources.ru.xlf
… src/Xamarin.Android.Build.Tasks/Properties/xlf/Resources.tr.xlf
… src/Xamarin.Android.Build.Tasks/Properties/xlf/Resources.zh-Hans.xlf
… src/Xamarin.Android.Build.Tasks/Properties/xlf/Resources.zh-Hant.xlf