@jonathanpeppers 2020-12-09
I was seeing an issue with SQLite NuGet packages, I think we'll need this addressed before we can finish this:

https://github.com/dotnet/sdk/issues/14908